+++ database/migrations/2016_07_10_202500_create_articles_table.php
@@ -14,10 +14,10 @@ class CreateArticlesTable extends Migration
     public function up()
     {
         Schema::create('articles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('collection_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('collection_id');
             $table->string('slug')->unique();
-            $table->unsignedInteger('position')->default(1);
+            $table->unsignedBigInteger('position')->default(1);
             $table->boolean('visible')->default(1);
             $table->string('image')->nullable();
             $table->timestamps();
@@ -26,8 +26,8 @@ class CreateArticlesTable extends Migration
         });
 
         Schema::create('article_languages', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('article_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('article_id');
             $table->char('language', 2);
             $table->string('title');
             $table->text('description')->nullable();
